+++ actix-http/tests/test_openssl.rs
@@ -275,7 +275,7 @@ async fn test_h2_head_binary() {
     let mut srv = test_server(move || {
         HttpService::build()
             .h2(|_| {
-                ok::<_, ()>(Response::Ok().content_length(STR.len() as u64).body(STR))
+                ok::<_, ()>(Response::Ok().body(STR))
             })
             .openssl(ssl_acceptor())
             .map_err(|_| ())